put sample cells and genes in the repo
